change id type to long

+++ Persistence/tadbir.Entities/InvoiceRow.cs
@@ -2,9 +2,9 @@ namespace tadbir.Entities
 {
     public class InvoiceRow
     {
-        public int InvoiceId { get; set; }
+        public long InvoiceId { get; set; }
         public Invoice Invoice { get; set; }
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
         public int Quantity { get; set; }
     }
